pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "evalforge"
version = "1.0.0"
description = "An OpenEnv RL environment where AI agents learn to evaluate LLM outputs for critical failure modes."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Team One9"},
]

dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["evalforge*"]